fix: overflow of device-name (#231)
## Issue
In the share-dialog long device-names might not be visible.

## Context
We had set `android:ellipsize="marquee"` in the layout but that only
works when the view is "selected".

## Solution
We now wrap the text to a max of two lines.
The bluetooth-beacon clips the device-name anyway. But in case the text
still overflows the start will be clipped to preserve the part that
likely identifies a device.